[tests] move witness util functions to blocktools.py
This avoids importing from segwit.py to bumpfee.py
